Move deploy and auth from generate to setup
