+++ packages/cli/src/commands/setup/auth/__tests__/auth.test.js
