+++ packages/cli/src/commands/setup/auth/providers/auth0.js
