+++ packages/cli/src/commands/setup/auth/providers/azureActiveDirectory.js
