+++ packages/cli/src/commands/setup/auth/providers/ethereum.js
