+++ packages/cli/src/commands/setup/auth/providers/firebase.js
