+++ packages/cli/src/commands/setup/auth/providers/goTrue.js
