+++ packages/cli/src/commands/setup/auth/providers/magicLink.js
