+++ packages/cli/src/commands/setup/auth/providers/supabase.js
