+++ packages/cli/src/commands/setup/auth/templates/auth.js.template
