+++ packages/cli/src/commands/setup/auth/templates/azureActiveDirectory.auth.js.template
