+++ packages/cli/src/commands/setup/auth/templates/ethereum.auth.js.template
