+++ packages/cli/src/commands/setup/auth/templates/firebase.auth.js.template
