+++ packages/cli/src/commands/setup/deploy/__tests__/deploy.test.js
